Split nfapi_vnf_interface.h into LTE, NR and common headers
Replaces the monolithic nfapi_vnf_interface.h with three files:
nfapi_lte_vnf_interface.h (LTE config and P5/P4/P7 functions),
nfapi_nr_vnf_interface.h (NR config and P5 functions, initially a
stub), and nfapi_vnf_interface_common.h (shared data types and P7
config). vnf_lte.h and vnf_common.h are updated to include the
appropriate new headers.
