+++ nfapi/oai_integration/aerial/fapi_vnf_p7.h
@@ -22,7 +22,7 @@
 
 #include "nfapi_interface.h"
 #include "nfapi_nr_interface_scf.h"
-#include "nfapi_vnf_interface.h"
+#include "nfapi_nr_vnf_interface.h"
 #include "fapi_nvIPC.h"
 #include "openair2/PHY_INTERFACE/queue_t.h"
 #include "vnf_p7_nr.h"
